Move get_PATH and x_subprocess_env functions into utils.py
