+++ powerline_shell/segments/bzr.py
@@ -1,25 +1,11 @@
-import os
 import subprocess
-from ..utils import RepoStats, ThreadedSegment
-
-
-def get_PATH():
-    """Normally gets the PATH from the OS. This function exists to enable
-    easily mocking the PATH in tests.
-    """
-    return os.getenv("PATH")
-
-
-def bzr_subprocess_env():
-    env = dict(os.environ)
-    env.update({"PATH": get_PATH()})
-    return env
+from ..utils import RepoStats, ThreadedSegment, get_subprocess_env
 
 
 def _get_bzr_branch():
     p = subprocess.Popen(['bzr', 'nick'],
                          stdout=subprocess.PIPE, stderr=subprocess.PIPE,
-                         env=bzr_subprocess_env())
+                         env=get_subprocess_env())
     branch = p.communicate()[0].decode("utf-8").rstrip('\n')
     return branch
 
@@ -49,7 +35,7 @@ def build_stats():
     try:
         p = subprocess.Popen(['bzr', 'status'],
                              stdout=subprocess.PIPE, stderr=subprocess.PIPE,
-                             env=bzr_subprocess_env())
+                             env=get_subprocess_env())
     except OSError:
         # Popen will throw an OSError if bzr is not found
         return (None, None)
