+++ powerline_shell/segments/svn.py
@@ -1,25 +1,12 @@
-import subprocess, os
-from ..utils import ThreadedSegment, RepoStats
-
-
-def get_PATH():
-    """Normally gets the PATH from the OS. This function exists to enable
-    easily mocking the PATH in tests.
-    """
-    return os.getenv("PATH")
-
-
-def svn_subprocess_env():
-    env = dict(os.environ)
-    env.update({"PATH": get_PATH()})
-    return env
+import subprocess
+from ..utils import ThreadedSegment, RepoStats, get_subprocess_env
 
 
 def _get_svn_revision():
     p = subprocess.Popen(["svn", "info", "--xml"],
                          stdout=subprocess.PIPE,
                          stderr=subprocess.PIPE,
-                         env=svn_subprocess_env())
+                         env=get_subprocess_env())
     for line in p.communicate()[0].decode("utf-8").splitlines():
         if "revision" in line:
             revision = line.split("=")[1].split('"')[1]
@@ -49,7 +36,7 @@ def build_stats():
     try:
         p = subprocess.Popen(['svn', 'status'],
                              stdout=subprocess.PIPE, stderr=subprocess.PIPE,
-                             env=svn_subprocess_env())
+                             env=get_subprocess_env())
     except OSError:
         # Popen will throw an OSError if svn is not found
         return None
