+++ test/segments_test/bzr_test.py
@@ -49,7 +49,7 @@ class BzrTest(unittest.TestCase):
         sh.bzr("branch", "trunk", branch)
         sh.cd(branch)
 
-    @mock.patch("powerline_shell.segments.bzr.get_PATH")
+    @mock.patch("powerline_shell.utils.get_PATH")
     def test_bzr_not_installed(self, get_PATH):
         get_PATH.return_value = "" # so bzr can't be found
         self.segment.start()
