+++ test/segments_test/hg_test.py
@@ -42,7 +42,7 @@ class HgTest(unittest.TestCase):
     def _checkout_new_branch(self, branch):
         sh.hg("branch", branch)
 
-    @mock.patch("powerline_shell.segments.hg.get_PATH")
+    @mock.patch("powerline_shell.utils.get_PATH")
     def test_hg_not_installed(self, get_PATH):
         get_PATH.return_value = "" # so hg can"t be found
         self.segment.start()
